PharmGKB curated drug list

A dataset containing all Chemicals with the Type of "Drug" in the PharmGKB knowledgebase. Not all of these drugs have been involved in PharmGKB annotations.

drug_pharmgkb

Format

A data frame with 3493 obs. of 24 variables:

pharmgkb_accession_id

Identifier assigned to this chemical by PharmGKB

name

Name PharmGKB uses for this drug

generic_names

Known generic names for this drug, comma-separated and "-enclosed

trade_names

Known trade names for this drug, comma-separated and "-enclosed

brand_mixtures

Known brand mixtures this drug is in, comma-separated and "-enclosed

type

Categories PharmGKB has assigned to this drug, can be more than one, possible values: Drug, Metabolite, Ion, Drug Class, Biological Intermediate, Small Molecule, Prodrug

cross_references

References to other resources in the form "resource:id", comma-separated

smiles

The SMILES structure for this drug

inchi

The InCHI key for this drug

dosing_guideline

"Yes" if PharmGKB has annotated a guideline with this drug, "No" otherwise

external_vocabulary

Term for this drug in another vocabulary in the form "vocabulary:id", comma-separated

clinical_annotation_count

Not described in drug.zip documentation

variant_annotation_count

Not described in drug.zip documentation

pathway_count

Not described in drug.zip documentation

vip_count

Not described in drug.zip documentation

dosing_guideline_sources

Not described in drug.zip documentation

top_clinical_annotation_level

Not described in drug.zip documentation

top_fda_label_testing_level

Not described in drug.zip documentation

top_any_drug_label_testing_level

Not described in drug.zip documentation

label_has_dosing_info

Not described in drug.zip documentation

has_rx_annotation

Not described in drug.zip documentation

rxnorm_identifiers

Not described in drug.zip documentation

atc_identifiers

Not described in drug.zip documentation

pubchem_compound_identifiers

Not described in drug.zip documentation

Multiple "Type" values can be assigned to a given entry since the same substance can be used in different contexts. The "Type" values used for Drugs & Chemicals are as follows:

Drug A chemical substance used in the treatment, cure,prevention, or diagnosis of disease.

Metabolite Any intermediate or product resulting from metabolism.

Ion An atomic or molecular particle having a net electric charge.

Drug Class A drug class is a group of medications that may work in the same way, have a similar chemical structure, or are used to treat the same health condition.

Biological Intermediate An endogenous small molucule or ion.

Small Molecule An electrically neutral entity consisting of more than one atom.

Prodrug A compound that must undergo chemical conversion by metabolic processes before becoming the pharmacologically active drug for which it is a prodrug.

Source

https://www.pharmgkb.org/downloads